Split code to more files
