+++ lib/screens/providers_list_screen.dart
@@ -4,8 +4,6 @@ All rights reserved.
 This source code is licensed under the BSD-style license found in the
 LICENSE file in the root directory of this source tree. */
 
-import 'dart:math' show Random;
-
 import 'package:flutter/material.dart';
 import 'package:lthn_vpn/models/exit_node_provider.dart';
 import 'package:provider/provider.dart';
