+++ lib/screens/vpn_instruction_screen.dart
@@ -4,100 +4,34 @@ All rights reserved.
 This source code is licensed under the BSD-style license found in the
 LICENSE file in the root directory of this source tree. */
 
-import 'dart:convert';
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:path/path.dart' as path;
-import 'package:path_provider/path_provider.dart';
-import 'package:permission_handler/permission_handler.dart';
-import 'package:ext_storage/ext_storage.dart';
-import 'package:flag/flag.dart';
 
 import '../providers/exit_node_providers.dart';
 import '../models/exit_node_service.dart';
-import '../widgets/service_tile_header.dart';
-import '../widgets/vpn_badge.dart';
+import '../widgets/vpn_instruction_header.dart';
+import '../widgets/save_ovpn_file_buton.dart';
 
 class VpnInstructionScreen extends StatelessWidget {
   static const routeName = '/vpn-instruction';
-
-  String _convertOvpnTemplate(String template, ExitNodeService _service) {
-    template = template.replaceAll('{ip}', '${_service.vpnEndpoint}');
-    template =
-        template.replaceAll('{port}', '${_service.vpnPort.split('/')[0]}');
-    template = template.replaceAll(
-        '{proto}', '${_service.vpnPort.split('/')[1].toLowerCase()}');
-    template = template.replaceAll('{tundev}', 'tun1');
-    template = template.replaceAll('{comment_dn}', '');
-    template = template.replaceAll('{tunnode}', '/dev/net/tun');
-    template = template.replaceAll('{mtu}', '1400');
-    template = template.replaceAll('{mssfix}', '1300');
-    template = template.replaceAll('{auth_file}', ''); // We could use this!
-    template = template.replaceAll('{pull_filters}', '');
-    template = template.replaceAll('{mgmt_comment}', '');
-    template = template.replaceAll('{mgmt_sock}', '127.0.0.1 11193');
-    template = template.replaceAll('{hproxy_comment}', '#');
-    template = template.replaceAll('{http_proxy}', '');
-    template = template.replaceAll('{http_proxy_port}', '');
-    template = template.replaceAll('{comment_syslog}', '');
-    template = template.replaceAll('{rgw_comment}', '');
-    template = template.replaceAll('{bdns_comment}', '#');
-    template = template.replaceAll('{f_ca}',
-        '${utf8.decode(base64.decode(_service.providerCertContent))}');
-    return template;
-  }
 
   @override
   Widget build(BuildContext context) {
     final ExitNodeService _service = ModalRoute.of(context).settings.arguments;
 
     return Consumer<ExitNodeProviders>(builder: (ctx, provider, child) {
-      final _exitNode =
-          provider.providers.firstWhere((p) => p.id == _service.providerId);
+      final _country = provider.providers
+          .firstWhere((p) => p.id == _service.providerId)
+          .country;
       return Scaffold(
         appBar: AppBar(
           title: Text('${_service.name}'),
         ),
         body: Column(
           children: <Widget>[
-            Card(
-              elevation: 4.0,
-              margin: EdgeInsets.all(6.0),
-              child: Column(
-                children: <Widget>[
-                  ListTile(
-                    leading: Flag(
-                      _exitNode.country.code,
-                      width: 60.0,
-                      height: 40.0,
-                      fit: BoxFit.cover,
-                    ),
-                    title: Text(
-                      _service.providerName,
-                      style: Theme.of(context).textTheme.headline6,
-                    ),
-                    subtitle: Text(
-                      '${_exitNode.country.name} ' +
-                          '${_exitNode.country.code}',
-                      style: Theme.of(context).textTheme.subtitle1,
-                    ),
-                    // trailing: const Icon(Icons.arrow_forward_ios),
-                    // onTap: () {
-                    //   Navigator.pushNamed(
-                    //     context,
-                    //     ServicesListScreen.routeName,
-                    //     arguments: _currentNode,
-                    //   );
-                    // },
-                  ),
-                  ServiceTileHeader(
-                    service: _service,
-                    badge: VpnBadge(),
-                  ),
-                ],
-              ),
+            VpnInstructionHeader(
+              country: _country,
+              service: _service,
             ),
             Expanded(
               child: Scrollbar(
@@ -125,56 +59,7 @@ class VpnInstructionScreen extends StatelessWidget {
                                     .primaryTextTheme
                                     .subtitle1,
                               ),
-                              CircleAvatar(
-                                backgroundColor: Theme.of(context).accentColor,
-                                child: IconButton(
-                                    padding: EdgeInsets.zero,
-                                    icon: Icon(Icons.save),
-                                    onPressed: () async {
-                                      var template = await DefaultAssetBundle
-                                              .of(context)
-                                          .loadString(
-                                              'assets/openvpn_client_template.txt');
-                                      template = _convertOvpnTemplate(
-                                          template, _service);
-                                      var status =
-                                          await Permission.storage.request();
-                                      final extDir = await ExtStorage
-                                          .getExternalStoragePublicDirectory(
-                                              ExtStorage.DIRECTORY_DOWNLOADS);
-                                      if (status == PermissionStatus.granted) {
-                                        var filePath = path.join(
-                                            extDir,
-                                            _service.providerName +
-                                                '_' +
-                                                _service.id +
-                                                '.ovpn');
-                                        filePath = filePath
-                                            .trim()
-                                            .replaceAll(' ', '_');
-                                        final ovpnFile = File(filePath);
-                                        await ovpnFile.writeAsString(template,
-                                            flush: true);
-                                        await showDialog(
-                                            context: context,
-                                            builder: (ctx) => AlertDialog(
-                                                  title: const Text(
-                                                      'OpenVPN configuration saved'),
-                                                  content: Text(
-                                                      'File path:\n${ovpnFile.path}'),
-                                                  actions: <Widget>[
-                                                    FlatButton(
-                                                      child: const Text('OK'),
-                                                      onPressed: () {
-                                                        Navigator.of(context)
-                                                            .pop();
-                                                      },
-                                                    ),
-                                                  ],
-                                                ));
-                                      }
-                                    }),
-                              ),
+                              SaveOvpnFileButton(service: _service),
                             ],
                           ),
                           const SizedBox(height: 14),
